+++ src/main/java/posojt/web/controllers/auth/LoginController.java
@@ -1,4 +1,4 @@
-package posojt.web.controllers;
+package posojt.web.controllers.auth;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -12,5 +12,6 @@ public class LoginController {
 		ModelAndView mv = new ModelAndView();
 		return mv;
 	}
+
 
 }
