+++ src/main/java/posojt/web/controllers/seller/SellerHomeController.java
@@ -1,18 +1,15 @@
-package posojt.web.controllers;
+package posojt.web.controllers.seller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.ModelAndView;
 
 @Controller
-public class UserController {
-	
-	
-	@RequestMapping("/user/home")
-	public ModelAndView userHome() {
-		ModelAndView mv = new ModelAndView("");
+public class SellerHomeController {
+
+	@RequestMapping("/seller/home")
+	public ModelAndView sellerHome() {
+		ModelAndView mv = new ModelAndView();
 		return mv;
 	}
-
-
 }
