+++ src/main/java/posojt/web/controllers/seller/ShopController.java
@@ -1,20 +1,14 @@
-package posojt.web.controllers;
+package posojt.web.controllers.seller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.ModelAndView;
 
 @Controller
-public class SellerController {
+public class ShopController {
 
-	@RequestMapping("/seller/home")
-	public ModelAndView sellerHome() {
-		ModelAndView mv = new ModelAndView();
-		return mv;
-	}
-	
 	@RequestMapping("/seller/shopList")
-	public ModelAndView adminUserList() {
+	public ModelAndView sellerUserList() {
 		ModelAndView mv = new ModelAndView("sellerShopList");
 		return mv;
 	}
