+++ src/main/java/posojt/web/controllers/user/UserHomeController.java
@@ -1,0 +1,16 @@
+package posojt.web.controllers.user;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.servlet.ModelAndView;
+
+@Controller
+public class UserHomeController {
+
+	@RequestMapping("/user/home")
+	public ModelAndView userHome() {
+		ModelAndView mv = new ModelAndView();
+		return mv;
+	}
+
+}
