+++ src/main/webapp/WEB-INF/layouts/adminHomeView.xml
@@ -3,11 +3,11 @@
 <tiles-definitions>
 	<definition name="adminUserList" extends="adminTemplate">
 		<put-attribute name="title" value="User list"/>
-		<put-attribute name="body" value="/WEB-INF/pages/admin/userList.jsp"></put-attribute>
+		<put-attribute name="body" value="/WEB-INF/pages/admin/user/userList.jsp"></put-attribute>
 	</definition>
 	<definition name="adminShopList" extends="adminTemplate">
 		<put-attribute name="title" value="User list"/>
-		<put-attribute name="body" value="/WEB-INF/pages/admin/shopList.jsp"></put-attribute>
+		<put-attribute name="body" value="/WEB-INF/pages/admin/seller/shopList.jsp"></put-attribute>
 	</definition>
 	
 </tiles-definitions>
