+++ src/main/webapp/WEB-INF/pages/common/adminTemplate.jsp
